+++ Transbank/Webpay/TransaccionCompletaMall/Common/DeferredPeriods.cs
@@ -8,12 +8,12 @@ namespace Transbank.Webpay.TransaccionCompletaMall.Common
     public class DeferredPeriods
     {
         [JsonProperty("amount")]
-        internal int Amount { get; }
+        internal decimal Amount { get; }
         
         [JsonProperty("period")]
         internal int Period { get; }
 
-        public DeferredPeriods(int amount, int period)
+        public DeferredPeriods(decimal amount, int period)
         {
             Amount= amount;
             Period = period;
